Release 0.40.0 (#322)
* fix: Add exception handler to AmityMQTT

* chore: Update version to 0.38.2

* Um/asc 20072/story deliverable two (#320)

* Added Story Live Object

* Added Support for comment and react on story

* feat: Reaction live collection (#310)

* Story targets done

- Target Live Collection
- Target Live Object
- Event Capturer added for the Story Click and Story View

* Added Analytics

* Added some enhancements for is State Seen

* Conflicts Resolved

* Commented AddTarget in StoryComposer because was causing error in live collection

---------


* Chore : Updated Version and ChangeLog.MD

* Chore : Added Gist Code for Deliverable two and make a correction in param name

---------

+++ lib/src/data/repo_impl/story_repo_impl.dart
@@ -1,28 +1,36 @@
+import 'dart:async';
 import 'dart:io';
 
 import 'package:amity_sdk/amity_sdk.dart';
 import 'package:amity_sdk/src/core/core.dart';
+import 'package:amity_sdk/src/core/mapper/story_model_mapper.dart';
 import 'package:amity_sdk/src/core/model/api_request/create_stroy_request.dart';
 import 'package:amity_sdk/src/core/model/api_request/story_delete_by_id_request.dart';
+import 'package:amity_sdk/src/core/utils/model_mapper.dart';
 import 'package:amity_sdk/src/data/converter/story/create_story_response_extension.dart';
 import 'package:amity_sdk/src/data/converter/story/story_hive_extension_converter.dart';
 import 'package:amity_sdk/src/data/data_source/data_source.dart';
+import 'package:amity_sdk/src/data/data_source/local/hive_entity/story_hive_entity_27.dart';
 import 'package:amity_sdk/src/data/data_source/remote/api_interface/story_api_interface.dart';
 import 'package:amity_sdk/src/domain/domain.dart';
+import 'package:amity_sdk/src/domain/repo/story_target_repo.dart';
 
 class StoryRepoImpl extends StoryRepo {
   final StoryApiInterface storyApiInterface;
   final DbAdapterRepo dbAdapterRepo;
   final FileRepo fileRepo;
+  final StoryTargetRepo storyTargetRepo;
 
   StoryRepoImpl(
       {required this.storyApiInterface,
       required this.dbAdapterRepo,
-      required this.fileRepo});
+      required this.fileRepo,
+      required this.storyTargetRepo});
 
   @override
   Future<AmityStory> createStory(CreateStoryRequest request) async {
     var entity = request.convertToHiveEntity();
+    // Update Story Target
     try {
       final AmityUploadResult<AmityFileInfo> amityUploadResult;
       final fileEntity = FileHiveEntity()
@@ -33,9 +41,14 @@ class StoryRepoImpl extends StoryRepo {
       dbAdapterRepo.fileDbAdapter.saveFileEntity(fileEntity);
 
       entity.syncState = AmityStorySyncState.CREATED.value;
+      entity.expiresAt = DateTime.now().add(Duration(days: 365));
       await dbAdapterRepo.storyDbAdapter.saveStoryEntity(entity);
+      storyTargetRepo.updateStoryTargetLocalLastStoryExpiresAt(
+          AmityStoryTargetTypeExtension.enumOf(request.targetType!),
+          request.targetId!,
+          entity.expiresAt!);
 
-      entity.syncState = AmityStorySyncState.UPLOADING.value;
+      entity.syncState = AmityStorySyncState.SYNCING.value;
       await dbAdapterRepo.storyDbAdapter.saveStoryEntity(entity);
 
       switch (AmityStoryDataTypeExtension.enumOf(request.dataType!)) {
@@ -62,7 +75,8 @@ class StoryRepoImpl extends StoryRepo {
           amityUploadResult = await fileRepo
               .uploadImage(UploadFileRequest(files: [File(request.uri!.path)])
                 ..uploadId = request.referenceId
-                ..fullImage = true).onError((error, stackTrace) {
+                ..fullImage = true)
+              .onError((error, stackTrace) {
             entity.syncState = AmityStorySyncState.FAILED.value;
             return Future.error(error!);
           });
@@ -88,6 +102,11 @@ class StoryRepoImpl extends StoryRepo {
     } catch (error) {
       entity.syncState = AmityStorySyncState.FAILED.value;
       dbAdapterRepo.storyDbAdapter.saveStoryEntity(entity);
+      entity = dbAdapterRepo.storyDbAdapter.getStoryEntity(entity.storyId!)!;
+      storyTargetRepo.updateStoryTargetLocalLastStoryExpiresAt(
+          AmityStoryTargetTypeExtension.enumOf(request.targetType!),
+          request.targetId!,
+          entity.expiresAt!);
       rethrow;
     }
   }
@@ -103,7 +122,7 @@ class StoryRepoImpl extends StoryRepo {
   @override
   Stream<List<AmityStory>> listenStories(
       RequestBuilder<GetStoriesByTragetRequest> request) {
-    return dbAdapterRepo.storyDbAdapter
+        Stream<List<AmityStory>> storyStream = dbAdapterRepo.storyDbAdapter
         .listenStoryEntities(request)
         .map((event) {
       final req = request.call();
@@ -120,6 +139,7 @@ class StoryRepoImpl extends StoryRepo {
 
       return list;
     });
+    return storyStream;
   }
 
   @override
@@ -129,8 +149,90 @@ class StoryRepoImpl extends StoryRepo {
 
   @override
   Future<bool> deleteStoryById(StoryDeleteByIdRequest params) async {
-    final data = await storyApiInterface.deleteStoryById(params);
+    bool data = false;
+    if (!params.storyId!.startsWith("LOCAL_")) {
+      try{
+        await storyApiInterface.deleteStoryById(params);
+        data = true;
+      }catch(e){
+        data = false;
+      }
+    }
     dbAdapterRepo.storyDbAdapter.deleteStoryEntity(params.storyId!);
     return data;
+  }
+
+  @override
+  Future<AmityStory?> fetchAndSave(String objectId) async {
+    var story = await getStoryById(objectId);
+    if (story != null) {
+      return story;
+    } else {
+      return Future.value(null);
+    }
+  }
+
+  @override
+  ModelMapper<StoryHiveEntity, AmityStory> mapper() {
+    return StoryModalMapper();
+  }
+
+  @override
+  StreamController<StoryHiveEntity> observeFromCache(String objectId) {
+    final streamController = StreamController<StoryHiveEntity>();
+    dbAdapterRepo.storyDbAdapter.listenPostEntity(objectId).listen((event) {
+      streamController.add(event);
+    });
+    return streamController;
+  }
+
+  @override
+  Future<StoryHiveEntity?> queryFromCache(String objectId) async {
+    return dbAdapterRepo.storyDbAdapter.getStoryEntity(objectId);
+  }
+
+  @override
+  Future<AmityStory> getStoryById(String storyId) async {
+    final data = await storyApiInterface.getStoryById(storyId);
+    final amityStory = await data.saveToDb<AmityStory>(dbAdapterRepo);
+    return amityStory[0];
+  }
+
+  @override
+  int getFailedStoriesCount(AmityStoryTargetType targetType, String targetId) {
+    return dbAdapterRepo.storyDbAdapter.getStoryCount(
+        targetType.value, targetId, [AmityStorySyncState.FAILED]);
+  }
+
+  @override
+  DateTime? getHighestStoryExpiresAt(
+      AmityStoryTargetType targetType, String targetId) {
+    return dbAdapterRepo.storyDbAdapter.getHighestStoryExpiresAt(
+        targetType.value,
+        targetId,
+        [AmityStorySyncState.SYNCING, AmityStorySyncState.SYNCED]);
+  }
+
+  @override
+  DateTime? getHighestSyncedStoryExpiresAt(
+      AmityStoryTargetType targetType, String targetId) {
+    return dbAdapterRepo.storyDbAdapter.getHighestStoryExpiresAt(
+        targetType.value, targetId, [AmityStorySyncState.SYNCED]);
+  }
+
+  @override
+  int getSyncingStoriesCount(AmityStoryTargetType targetType, String targetId) {
+    return dbAdapterRepo.storyDbAdapter.getStoryCount(targetType.value,
+        targetId, [AmityStorySyncState.SYNCING, AmityStorySyncState.UPLOADING]);
+  }
+  
+  @override
+  List<AmityStory> getFailedStories(String targetType, String targetId) {
+    List<AmityStory> stories = [];
+    var entites = dbAdapterRepo.storyDbAdapter.getStoriesBySyncStates(targetType, targetId, [AmityStorySyncState.FAILED]);
+    for (var entity in entites) {
+      stories.add(entity.convertToAmityStory());
+    }
+    return stories;
   }
 }
